backend/requirements.txt
strands-agents==1.0.1
bedrock-agentcore==0.1.0
awslabs.aws-api-mcp-server